+++ tests/test_transport.py
@@ -1,0 +1,87 @@
+"""Tests for Transport Utility Classes"""
+
+import pytest
+
+
+class TestValidRequestObject:
+    """Tests for ValidRequestObject class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test that a ValidRequestObject instance can be initialized"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_bool(self):
+        """Test boolean value of a ValidRequestObject"""
+
+
+class TestInvalidRequestObject:
+    """Tests for InvalidRequestObject class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test that a ValidRequestObject instance can be initialized"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_bool(self):
+        """Test boolean value of a ValidRequestObject"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_add_error(self):
+        """Test that an error can be added"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_has_errors(self):
+        """Test that errors are correctly returned/reported"""
+
+
+class TestResponseObject:
+    """Tests for ResponseObject class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test that a ResponseObject instance can be initialized"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_bool(self):
+        """Test boolean value of a ResponseObject"""
+
+
+class TestResponseSuccessCreated:
+    """Tests for ResponseSuccessCreated class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test that a ResponseSuccessCreated instance can be initialized"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_bool(self):
+        """Test boolean value of a ResponseSuccessCreated"""
+
+
+class TestResponseSuccessWithNoContent:
+    """Tests for ResponseSuccessWithNoContent class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test that a ResponseSuccessWithNoContent instance can be initialized"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_bool(self):
+        """Test boolean value of a ResponseSuccessWithNoContent"""
+
+
+class TestResponseFailure:
+    """Tests for ResponseFailure class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test that a ResponseFailure instance can be initialized"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_bool(self):
+        """Test boolean value of a ResponseFailure"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_value(self):
+        """Test retrieval of ResponseFailure information"""