+++ tests/test_usecase.py
@@ -1,0 +1,91 @@
+"""Tests for Usecase Functionality"""
+
+import pytest
+
+
+class TestUseCase:
+    """Tests for the generic UseCase Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test Initialization of the generic UseCase class"""
+
+
+class TestShowRequestObject:
+    """Tests for the generic ShowRequest Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test Initialization of the generic ShowRequest class"""
+
+
+class TestListRequestObject:
+    """Tests for the generic ListRequest Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test Initialization of the generic ListRequest class"""
+
+
+class TestCreateRequestObject:
+    """Tests for the generic CreateRequest Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test Initialization of the generic CreateRequest class"""
+
+
+class TestUpdateRequestObject:
+    """Tests for the generic UpdateRequest Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test Initialization of the generic UpdateRequest class"""
+
+
+class TestDeleteRequestObject:
+    """Tests for the generic DeleteRequest Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_init(self):
+        """Test Initialization of the generic DeleteRequest class"""
+
+
+class TestShowUseCase:
+    """Tests for the generic ShowUseCase Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_process_request(self):
+        """Test Show UseCase's `process_request` method"""
+
+
+class TestListUseCase:
+    """Tests for the generic ListUseCase Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_process_request(self):
+        """Test List UseCase's `process_request` method"""
+
+
+class TestCreateUseCase:
+    """Tests for the generic CreateUseCase Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_process_request(self):
+        """Test Create UseCase's `process_request` method"""
+
+
+class TestUpdateUseCase:
+    """Tests for the generic UpdateUseCase Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_process_request(self):
+        """Test Update UseCase's `process_request` method"""
+
+
+class TestDeleteUseCase:
+    """Tests for the generic DeleteUseCase Class"""
+
+    @pytest.mark.skip(reason="To Be Implemented")
+    def test_process_request(self):
+        """Test Delete UseCase's `process_request` method"""