+++ CHANGELOG.rst
@@ -17,3 +17,9 @@ Changelog
 
 * Add `bleach` as a setup requirement
 * Add GeoPoint and Decimal Data Types to Entities
+
+0.0.4 (2018-)
+------------------
+
+* UseCase Utility Classes
+* Repository Abstract Classes